+++ doc/Changelog.md
@@ -2,10 +2,14 @@
 
 # git master
 
-* [290](https://github.com/Eyescale/Lunchbox/pull/290):
-  A ThreadPool class
+* [294](https://github.com/Eyescale/Lunchbox/pull/294):
+  * Add description to plugins
+  * Add lunchbox::string::prepend()
+  * Add lunchbox::term::getSize()
 * [292](https://github.com/Eyescale/Lunchbox/pull/292):
   Add MemoryMap element getter
+* [290](https://github.com/Eyescale/Lunchbox/pull/290):
+  A ThreadPool class
 
 # Release 1.15 (09-12-2016)